Changed notation regarding DG
Now, it is not dG0, dG1, dG2, instead we indicate the number of local degrees of freedom.
dG1 corresponds to dG0, dG3 to dG1 and dG6 to dG2.

New is dG8: this is the gradient space of CG2, such that the gradient of the velocity can
be exactly represented in the stress/strain space.

Code Generation:
Startet to clean this up. Uses the new dG-notation. Further, python scripts are now called
as the generated hpp-Files. A new run.sh script is provided to reproduce all hpp-files.