Strict capping of damage
Here I cap damage at 1.-1e-12, effectively demanding that damage always be < 1. This is necessary to prevent a NaN from appearing in ``multiplicator``. The time evolution equations don't allow damage to become equal to one - only to approach it asymptotically. I guess what's happening is that there's a slight overshoot in the advection (as is to be expected) and that this is only capped to ``<= 1`` and not ``< 1``.