A minor bug fix and some cosmetic changes to the 0-layer model
I had some problems understanding the results, so I went through this with a fine-toothed comb. It's fine in the end (I'm just using too long time steps) - but I found that adding a bunch of 'const' helps to make sure. Also, found a bug where the ice temperature is set in K, and not C when everything melts (not sure if it matters, but it gave surprising results).