Implement albedo modifications for the zero layer model
Here I fix the implementation of the zero layer model to follow Semtner's original formulation in modifying the albedo using the extinction coefficient I0. This fixes the original formulation in neXtSIM which was lacking in this respect. This commit also includes some minor bug fixes. Everything seems to work fine now.