Use free-drift solution in open water
Here I change the existing limiting in ``prepareIteration``, so that we
assume that open water is covered by very thin ice (A = H = 1e-4). This
in essence enforces a free-drift solution everywhere there is very thin
or no ice. It fixes the crash I got with mEVP, but strangely enough the
"window" I advect is smeared towards the leading edge towards the end of
the simulation. Investigation is on-going.