New implementation of the explicit solver for BBM
I've re-written the explicit solver in the MEB/BBM portion of the code. This follows the BBM paper, which is essentially a copy of Hunke & Dukowicz's equations (I just added minor stabilisation). The free-drift case seems to work well, but I haven't tested the rest properly. There are several outstanding items labelled with a "FIXME".