Further improvements of the initial conditions for free drift
I realised that using H = 1 m in the ice patch actually results in a
slab thickness of 10 m (if A = 0.1), and that doesn't compare well with
"classical free drift", where we assume slab thickness must be << 1 m. So
I changed this. I also updated the mEVP free drift initial conditions to
fit with the BBM case.