Change in commenting style for long-block commenting
As Piotr and I figure out which implementation of MEBStep is better we are simply commenting them out when testing. We were using C-style commenting from this (i.e., /* ... */), but this conflicted with other C-style comments in the code. Here I use CPP commenting (i.e., #if 0 ... #endif), which we should only do in exceptional cases ... but works well here.