Use Véro's definition of damage
I'm changing the definition of damage, so that d=1 us undamaged and d=0
is fully damaged. This is the opposite of Pierre and Sylvain's
definition, that we've used so far. The reason for the change is that
the model is most sensitive to changes in damage close to the fully
damaged state, and advecting a small number (\epsilon) accurately is
easier than advecting 1-\epsilon.