Two bug fixes for the Winton thermodynamics
The first bug fix moves .resize() calls from the constructor to the
setData method and adds a call to IIceThermodynamcis::setData to the
ThermoWinton::setData method. This ensures that all data are properly
initialised, as calling .resize() in the constructor is too early.

The second bug fix is a Celsius vs Kelvin error in the melting point
temperature.