Prevent crash at the end
A somewhat strange fix to the model crashing at the end of the run on
some configurations. Instead of using find and erasing map entries in
timeIndexByFile using the iterator I use the key directly instead (also
for openFiles, to be consistent). The two should be functionally the
same, but using the iterator crashes while using the key does not.