Implicitly assume forcing is DG order 1
When interpolating the wind and currents from the forcing (from file or
coupler) we should assume this is constant over the element - i.e. set
DG = 1 in the input to DG2CG. This was a higher order before, which
introduced random values into the resulting CG vector.

The model now runs when thermodynamics are commented out - but turning
thermo on crashes it.