WIP fix build for gfortran
I have manually changed lib ifcore to lgfortran. This will need to fixed
properly for intel and gnu compiler and not hard-coded.

this is a temporary fix.

The order of libs is also incorrect for XIOS_LIBRARY_LIST. xios depends
on netcdf so it should come second.

There was an erroneous ::Xios for the constructor when creating
xios_handler.