disable non-MPI tests when building with MPI
disabled the following tests when MPI is enabled:
* testRectGrid
* testParaGrid
* testConfigOutput

The latter two have only been disabled temporarily.

I have raised issue #454 explaining why the tests are disabled.

They will be re-enabled after paragrid is parallelized with MPI.

`testRectGrid` has a serial version and an MPI version
(`testRectGrid_MPI`). We run the MPI version when MPI is enabled and the
serial otherwise.