A fix to the crash reported in issue 470
The function Logged::configure() is trying to access keyMap, but this is
not accessible. I propose fixing this by adding keyMap to Logged.hpp in
the protected clause. I'm not sure if this is the right place, or if
it should be public or private. But the model runs now (at least the
thermo test).