Change the "expected dimensions of the benchmark domain"
In the BenchmarkBoundaries_test.cpp file I change the nx and ny numbers
to fit with what we expect from a typical benchmark run. This
coincidently fixes the fact that the test doesn't run after me updating
BenchmarkCoordinates.hpp in a previous commit.

A better fix would be to put a doctest::Approx into the the
"REQUIRE" statement ... but that's not needed for now.