Update the mesh boundary handling:
** periodic boundary

a periodic boundary element consists of
- an edge (given by its id)
- two elements on the left/right or bottom/top of the edge
- two indices specifying where the edge is as seen from the element

** Dirichlet-type conditions:

- dirichlet (v=0)
- open (free inflow/outflow with zero concentration on inflow)
- freeslip (normal velocity zero)

the all have the interface std::array<std::vector<size_t>, 4>
where the outer index corresponds to elements that have boundary on the bottom, right, top, left
and the inner index to corresponding id of the element.

*** Changes
Handling of periodic boundaries required several changes in DGTransport.cpp and cgParametricMomentum.cpp
This mostly includes restructuring code and adding the edge (bottom, right, top, left) as template
parameter.

Further, code is added to deal with open and freeslip boundaries.