Make sure damage is ``resize``d in the healing module
The Shared::DAMAGE variable was not being resized as part of the
IceGrowth test, so the test threw a segfault. It's easily fixed by
adding ``damage.data().resize`` to the IDamageHealing.hpp ``setData``
function. This is not problematic for the full model, but I don't
understand why it too didn't throw a segfault.