Get around the "deleted function" error
When compiling with the Intel compiler we get an error that "the default
constructor of "Nextsim::DynamicsKernel<6, 8>" cannot be referenced --
it is a deleted function". The solution is not to use the "= default"
keyword, but to manually define a constructor which sets the key
parameters to sensible default values. It is not clear why the other
compilers don't flag this.