Ensure that linked issues close
This should (hopefully) ensure that linked issues close automatically
when the PR is merged. The trick is to have one of the keywords close,
closes, closed, fix, fixes, fixed, resolve, resolves, or resolved
followed by #ISSUE_NUMBER. In the original PR template the heading was
Fixes \#(your issue number), so as long as the "/" was there github
didn't pick up on it.