Implement OASIS calls
Move from all comments to actual code that compiles for the OASIS calls.
We still don't have any passing of information to the ocean - this
needs implementing. Most of the fields from the ocean to ice are
implemented already, but we're missing ssh and we need special handling
for mld (which is optional).