Move common OASIS calls out of OASISCoupled class
The OASIS calls, init_comp, get_localcomm, terminate, def_partition, and
end_def can only be called once, so they must be moved out of the
OASISCoupled class and it's children. This makes the program structure
less nice, but I can't think of a better way to do it. It also leaves
OASISCoupled with only OASISTime and updateOASISTime(), but I suppose
that it's worth having anyway.