Add a function to check fields to ModelComponent
The function checkFields will check all fields listed in fieldsToCheck,
using overElements and checkFieldsElement (which does all the work).
checkFieldsElement checks if the given fields are within the bounds
listed in fieldsToCheck and throws a runtime_error if the check fails.
Each model component can then populate it's own fieldsToCheck vector and
call checkFields to check it's internal logic.