First step in simplifying thermodynamics data management
This is to address issue #725. I removed all duplication of prognostic
variables in the model and also removed the slab ice and slab snow
thickness variables. The assumption will be that all shared variables
are averaged over the grid cell (as is standard), and routines that need
the slab thickness calculate this on-the-fly.

TODO: Figure out if the tests run (some will not and will need
modification. Check if getState and friends are still doing sensible
things. Check if we have dangling, unused definitions lying around.