Handle exception thrown by checkFieldsElement
This adds a try-catch clause in iterator, which then exits the main
model loop with a non-zero return code, printing the boost-generated
stack trace to the command line. This way, the model exits gracefully,
closing all files and creating a restart file, but returns a non-zero
return code.

The only problem here is that boost stacktrace may not always be
implemented, so the code may not compile on all machines. We need to
discuss how to address this, because the stack trace is a very usefull
thing to have.

It would also be nice to have the i and j index and not only the linear
index. And it would be nice to have the file name and line number, not
just "in [...] libnextsimlib.so".