Limit damage after dynamics
Otherwise, we get values like 1+1e-9 > 1, which is no good. It seems to
come from the projection from Gauss points back to DG space. Ideally,
we'd like to limit only once, but I'm not sure if that's possible.

I also changed avgU.zero() to avgU.setZero(), although I'm not sure what
the practical difference is.