Two minor bug fixes to the physics
These became apparent when pushing the model using the field-checking
implemented in this branch, so I'm including them here. One sets qio to
zero if there's no ice and the other sets the tice to the freezing
point of ice when there's no ice in ThermoIce0. In the Winton model,
tice was already correctly set.