Put actual data in the oasis put calls
Populate the oasis put calls with data that's now available in the
model. I also added the ability to take the square root of a ModelArray,
and a test of this. I needed this for NEMO's requirement for the
magnitude of the ocean surface stress.

I didn't use m_couplingArrays and I'm not sure how to do that well.
There's a lot of repeat code here and this should be reduced.