Use setFieldsToCheck in ERA5 and TOPAZ boundary classes
Instead of a manual push_back of field names and bounds, use the
functionality in setFieldsToCheck. This is possible, since all the
fields these routines should check are registered in the
ModelArrayReference store.

To get this to work, I had to move the check
for empty fields from setFieldsToCheck to checkFields, since it's not
guarantied anymore that all fields have been set/resized when
setFieldsToCheck is called. I also moved several static constants in
ERA5Atmosphere.cpp and TOPAZOcean.cpp into the respective name spaces to
avoid conflict with constants named in PrognosticData.cpp. Finally, I
moved the call to setFieldsToCheck into the configure() function of the
respective classes. Logically, it belongs there, and it can now be
called from there as well, because it doesn't check if the fields have
been set/resized.