Use try-catch throughout and not return codes
See Tim's comment on PR #784. Now Iterator::run() catches the exception,
calls iterant.stop(t) and rethrows the exception. This is then caught by
Model::run(), which catches the exception, calls writeRestartFile() and
Finalizer::finalize(), and rethrows the exception. This is caught by
main, which logs the error using Logged::error, sets a return code and
continues as before.

I also changed Logged output defaults, so that Logged::error generates
output for both the log file and stdout/stderr.

We could return different codes for different types of errors, but this
is not implemented, and maybe not needed either.