Add tests to ConfigOutput_test.cpp to check snapshots and averages
The new tests use a common function to write output files with either
snapshotting or not. They then check if the output is properly averaged.
The tests work when run seperately, but the whole suite fails in varying
ways. Sometimes it's a crash and sometimes the second test fails
because the numbers don't add up. This has probably something to do with
my re-use of Configurator.