New CheckingModelComponent design
Mostly following Tim's review suggestions. The check itself now takes
place in ModelArray. Construction of a model array may initialise the
physically reasonable limits, but these can also be set via a setLimits
function. CheckingModelComponent now only provides booleans and a list
of variables to check, as well as a minimal checking function for
modules to call. Implementation examples are in IOceanBoundary and
TOPAZOcean, as well as IAtmosphereBoundary and ERA5Atmosphere. I
implemented a "global" check for PrognosticData, where all the variables
in the ModelArrayDataStore are checked. The check is more useful if the
modules have set the physical limits of their variables (otherwise these
are set to +/- infinity). But we always check for NaNs. The model fails
these tests because there are some uninitialised ModelArrays hanging
around.