Switch to the new batch of TOPAZ4 files
There's a bit of history here. The short version is that I've had no end
of trouble with the old TOPAZ4 files. This culminated in this branch,
because the script would crash in an unpredictable fashion (at least
for me), showing fields that were full of nonsensical values. However,
Tim W. has produced a new batch of TOPAZ files, which is what we're
using for Lagrangian neXtSIM these days. So, I decided to try these
instead, and now everything works as expected.