Add melt flux to qow, not qio
We should modify qow and not qio when all the ice melts away. This is
because when ciceDG becomes zero, the ice-portion is lost, as the total
flux is a multiplication of ciceDG and ice-only fluxes.

Also sneaks in a small fix to the dynamics, where we forgot to send the
snow to the dynamical core in the mEVP implementation.