Remove fine-grained control on checking
Removes the option to check or not to check in individual modules. This
is not needed in general. Now, only PrognosticData decides if we do a
fast check (debug.check_fields_fast, true by default), or if everyone
checks everything (debug.check_fields, false by default). I also renamed
the boolean checking variables to reflect this change.