Don't compare a float using ==
To check if the first or last row contains the pole we used "== 90",
which could, strictly speaking, return a false even when it should be
true if the value is very close to 90, but not 90, due to machine
imprecision. So, knowing that the ERA5 latitudes are spaced at 0.25
degrees, I changed this to "> 89.8". The fields produced look the same
after the change (as they should.