Remove ModelArrayReferenceStore::getArrayRef and code using it
It's not a safe function because the the storeRW.at(field) or
storeRO.at (field) may not have been initialised (see Tim's comment on
github). I also removed the check for velocities in PrognosticData when
debug.check_fields_fast is true, because this depended on getArrayRef.
Removing this check is ok, because errors in the velocity field will
quickly propagate to the concentration and thickness fields, which are
still checked.