Revert to a simpler version of ConfigOutput_test.cpp
I went back to the version that's in develop and only test to see if the
averaging works (it does). Some manual changes to the test show that the
snapshots also work. The problem arises when I try to test the two cases
in the same test file, but this has nothing to do with averaging etc,
it's probably a problem with file handling and the testing structure,
which I don't understand.